More fixes of benign warnings in indexer server (#776)
* Unused import

* Fix shadowing warnings

* Too polymorphic

* Move orphans for types from deps to class decl

* Drop unused erro messafge

* Unused binding

* Shadowing

* Too polymorphic warning

+++ index-server/src/Ergvein/Index/Server/BlockchainScanning/Common.hs
@@ -132,7 +132,7 @@ isEnoughSpace = do
   setGauge availableSpaceGauge $ fromIntegral (availSpace - requiredAvailSpace)
   pure $ requiredAvailSpace <= availSpace
  where
-  requiredAvailSpace = 2^30 -- 1Gb
+  requiredAvailSpace = 2^(30::Int) -- 1Gb
 
 feesScanning :: ServerM [Thread]
 feesScanning = sequenceA
